+++ src/features/account/components/details/Details.tsx
@@ -1,5 +1,6 @@
 import { useAuthContext } from '@contexts/AuthContext';
 
+import Name from './Name';
 import Email from './Email';
 
 export default function Details() {
@@ -7,6 +8,7 @@ export default function Details() {
 
   return (
     <div className="flex flex-col gap-y-10">
+      <Name />
       <Email user={user} />
     </div>
   );